requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.0
chromadb==0.4.15
sqlalchemy==2.0.23
python-dotenv==1.0.0
sentence-transformers==2.2.2
numpy==1.24.3
pydantic-settings==2.0.3
requests==2.31.0
slowapi==0.1.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

